+++ packages/flame/lib/src/components/component.dart
@@ -173,6 +173,20 @@ class Component with Loadable {
     nextParent = component;
   }
 
+  final List<Component> _ancestors = [];
+
+  /// A list containing the current parent and its parent, and so on, until it
+  /// reaches a component without a parent.
+  List<Component> ancestors() {
+    _ancestors.clear();
+    for (var currentParent = parent;
+        currentParent != null;
+        currentParent = currentParent.parent) {
+      _ancestors.add(currentParent);
+    }
+    return _ancestors;
+  }
+
   /// It receives the new game size.
   /// Executed right after the component is attached to a game and right before
   /// [onLoad] is called.
@@ -318,7 +332,7 @@ class Component with Loadable {
         parentGame.hasLayout,
         '"prepare/add" called before the game is ready. '
         'Did you try to access it on the Game constructor? '
-        'Use the "onLoad" or "onParentMethod" method instead.',
+        'Use the "onLoad" or "onMount" method instead.',
       );
       if (parentGame is FlameGame) {
         parentGame.prepareComponent(this);