+++ packages/flame/CHANGELOG.md
@@ -23,6 +23,8 @@
  - Remove `super.render` calls that are no longer needed
  - Fixed typo in error message
  - `TextPaint` to use `TextStyle` (material) instead of `TextPaintConfig`
+ - Underlying `Shape`s in `ShapeComponent` transform with components position, size and angle
+ - `HitboxShape` takes parents ancestors transformations into consideration (not scaling)
 
 ## [1.0.0-releasecandidate.16]
  - `changePriority` no longer breaks game loop iteration